# Speedup nuclei detection on small ROIs