--- a/server/ComputeNucleiFeatures/ComputeNucleiFeatures.py
+++ b/server/ComputeNucleiFeatures/ComputeNucleiFeatures.py
@@ -103,7 +103,7 @@ def main(args):
 
     print('\n>> CLI Parameters ...\n')
 
-    print args
+    print(args)
 
     check_args(args)
 
@@ -119,9 +119,14 @@ def main(args):
     #
     print('\n>> Creating Dask client ...\n')
 
+    start_time = time.time()
+
     c = cli_utils.create_dask_client(args)
 
-    print c
+    print(c)
+
+    dask_setup_time = time.time() - start_time
+    print('Dask setup time = {} seconds'.format(dask_setup_time))
 
     #
     # Read Input Image
@@ -132,19 +137,26 @@ def main(args):
 
     ts_metadata = ts.getMetadata()
 
-    print json.dumps(ts_metadata, indent=2)
+    print(json.dumps(ts_metadata, indent=2))
 
     is_wsi = ts_metadata['magnification'] is not None
 
     #
     # Compute tissue/foreground mask at low-res for whole slide images
     #
-    if is_wsi:
+    if is_wsi and process_whole_image:
 
         print('\n>> Computing tissue/foreground mask at low-res ...\n')
 
+        start_time = time.time()
+
         im_fgnd_mask_lres, fgnd_seg_scale = \
             cli_utils.segment_wsi_foreground_at_low_res(ts)
+
+        fgnd_time = time.time() - start_time
+
+        print('low-res foreground mask computation time = {}'.format(
+            cli_utils.disp_time_hms(fgnd_time)))
 
     #
     # Compute foreground fraction of tiles in parallel using Dask
@@ -174,12 +186,18 @@ def main(args):
 
         num_tiles = ts.getSingleTile(**it_kwargs)['iterator_range']['position']
 
-        print 'Number of tiles = %d' % num_tiles
+        print('Number of tiles = {}'.format(num_tiles))
 
-        tile_fgnd_frac_list = htk_utils.compute_tile_foreground_fraction(
-            args.inputImageFile, im_fgnd_mask_lres, fgnd_seg_scale,
-            **it_kwargs
-        )
+        if process_whole_image:
+
+            tile_fgnd_frac_list = htk_utils.compute_tile_foreground_fraction(
+                args.inputImageFile, im_fgnd_mask_lres, fgnd_seg_scale,
+                **it_kwargs
+            )
+
+        else:
+
+            tile_fgnd_frac_list = [1.0] * num_tiles
 
         num_fgnd_tiles = np.count_nonzero(
             tile_fgnd_frac_list >= args.min_fgnd_frac)
@@ -188,10 +206,11 @@ def main(args):
 
         fgnd_frac_comp_time = time.time() - start_time
 
-        print 'Number of foreground tiles = %d (%.2f%%)' % (
-            num_fgnd_tiles, percent_fgnd_tiles)
+        print('Number of foreground tiles = {0:d} ((1:2f)%%)'.format(
+            num_fgnd_tiles, percent_fgnd_tiles))
 
-        print 'Time taken = %s' % cli_utils.disp_time_hms(fgnd_frac_comp_time)
+        print('Tile foreground fraction computation time = {}'.format(
+            cli_utils.disp_time_hms(fgnd_frac_comp_time)))
 
     #
     # Detect and compute nuclei features in parallel using Dask
@@ -229,8 +248,9 @@ def main(args):
 
     nuclei_detection_time = time.time() - start_time
 
-    print 'Number of nuclei = ', len(nuclei_annot_list)
-    print "Time taken = %s" % cli_utils.disp_time_hms(nuclei_detection_time)
+    print('Number of nuclei = {}'.format(len(nuclei_annot_list)))
+    print('Nuclei detection time = {}'.format(
+        cli_utils.disp_time_hms(nuclei_detection_time)))
 
     #
     # Write annotation file
@@ -268,7 +288,8 @@ def main(args):
 
     total_time_taken = time.time() - total_start_time
 
-    print 'Total analysis time = %s' % cli_utils.disp_time_hms(total_time_taken)
+    print('Total analysis time = {}'.format(
+        cli_utils.disp_time_hms(total_time_taken)))
 
 
 if __name__ == "__main__":