--- a/server/ComputeNucleiFeatures/ComputeNucleiFeatures.py
+++ b/server/ComputeNucleiFeatures/ComputeNucleiFeatures.py
@@ -23,7 +23,8 @@ sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), '..')))
 from cli_common import utils as cli_utils  # noqa
 
 
-def compute_tile_nuclei_features(slide_path, tile_position, args, **it_kwargs):
+def compute_tile_nuclei_features(slide_path, tile_position, args, it_kwargs,
+                                 src_mu_lab=None, src_sigma_lab=None):
 
     # get slide tile source
     ts = large_image.getTileSource(slide_path)
@@ -40,7 +41,9 @@ def compute_tile_nuclei_features(slide_path, tile_position, args, **it_kwargs):
     # perform color normalization
     im_nmzd = htk_cnorm.reinhard(im_tile,
                                  args.reference_mu_lab,
-                                 args.reference_std_lab)
+                                 args.reference_std_lab,
+                                 src_mu=src_mu_lab,
+                                 src_sigma=src_sigma_lab)
 
     # perform color decovolution
     w = cli_utils.get_stain_matrix(args)
@@ -213,6 +216,26 @@ def main(args):
             cli_utils.disp_time_hms(fgnd_frac_comp_time)))
 
     #
+    # Compute reinhard stats for color normalization
+    #
+    src_mu_lab = None
+    src_sigma_lab = None
+
+    if is_wsi and process_whole_image:
+
+        print('\n>> Computing reinhard color normalization stats ...\n')
+
+        start_time = time.time()
+
+        src_mu_lab, src_sigma_lab = htk_cnorm.reinhard_stats(
+            args.inputImageFile, 0.01, magnification=args.analysis_mag)
+
+        rstats_time = time.time() - start_time
+
+        print('Reinhard stats computation time = {}'.format(
+            cli_utils.disp_time_hms(rstats_time)))
+
+    #
     # Detect and compute nuclei features in parallel using Dask
     #
     print('\n>> Detecting nuclei and computing features ...\n')
@@ -232,7 +255,9 @@ def main(args):
         cur_result = dask.delayed(compute_tile_nuclei_features)(
             args.inputImageFile,
             tile_position,
-            args, **it_kwargs)
+            args, it_kwargs,
+            src_mu_lab, src_sigma_lab
+        )
 
         # append result to list
         tile_result_list.append(cur_result)