--- a/server/NucleiDetection/NucleiDetection.py
+++ b/server/NucleiDetection/NucleiDetection.py
@@ -1,27 +1,17 @@
-import time
-import_time = time.time()
-
 import os
 import sys
 import json
+import time
 import itertools
 
 import numpy as np
 import dask
 
-htk_import_time = time.time()
-
 import histomicstk.preprocessing.color_normalization as htk_cnorm
 import histomicstk.preprocessing.color_deconvolution as htk_cdeconv
 import histomicstk.utils as htk_utils
 
-htk_import_time = time.time() - htk_import_time
-
-large_import_time = time.time()
-
 import large_image
-
-large_import_time = time.time() - large_import_time
 
 from ctk_cli import CLIArgumentParser
 
@@ -30,8 +20,6 @@ logging.basicConfig(level=logging.CRITICAL)
 
 sys.path.append(os.path.normpath(os.path.join(os.path.dirname(__file__), '..')))
 from cli_common import utils as cli_utils  # noqa
-
-import_time = time.time() - import_time
 
 
 def detect_tile_nuclei(slide_path, tile_position, args, **it_kwargs):
@@ -72,15 +60,6 @@ def detect_tile_nuclei(slide_path, tile_position, args, **it_kwargs):
 
 def main(args):
 
-    print('histomicsk import time = {}'.format(cli_utils.disp_time_hms(
-        htk_import_time)))
-
-    print('large_image import time = {}'.format(cli_utils.disp_time_hms(
-        time.time() - large_import_time)))
-
-    print('Total import time = {}'.format(cli_utils.disp_time_hms(
-        time.time() - import_time)))
-
     total_start_time = time.time()
 
     print('\n>> CLI Parameters ...\n')
@@ -109,9 +88,15 @@ def main(args):
     #
     print('\n>> Creating Dask client ...\n')
 
+    start_time = time.time()
+
     c = cli_utils.create_dask_client(args)
 
     print(c)
+
+    dask_setup_time = time.time() - start_time
+    print('Dask setup time = {}'.format(
+        cli_utils.disp_time_hms(dask_setup_time)))
 
     #
     # Read Input Image
@@ -129,12 +114,19 @@ def main(args):
     #
     # Compute tissue/foreground mask at low-res for whole slide images
     #
-    if is_wsi:
+    if is_wsi and process_whole_image:
 
         print('\n>> Computing tissue/foreground mask at low-res ...\n')
 
+        start_time = time.time()
+
         im_fgnd_mask_lres, fgnd_seg_scale = \
             cli_utils.segment_wsi_foreground_at_low_res(ts)
+
+        fgnd_time = time.time() - start_time
+
+        print('low-res foreground mask computation time = {}'.format(
+            cli_utils.disp_time_hms(fgnd_time)))
 
     #
     # Compute foreground fraction of tiles in parallel using Dask
@@ -166,10 +158,16 @@ def main(args):
 
         print('Number of tiles = {}'.format(num_tiles))
 
-        tile_fgnd_frac_list = htk_utils.compute_tile_foreground_fraction(
-            args.inputImageFile, im_fgnd_mask_lres, fgnd_seg_scale,
-            **it_kwargs
-        )
+        if process_whole_image:
+
+            tile_fgnd_frac_list = htk_utils.compute_tile_foreground_fraction(
+                args.inputImageFile, im_fgnd_mask_lres, fgnd_seg_scale,
+                **it_kwargs
+            )
+
+        else:
+
+            tile_fgnd_frac_list = [1.0] * num_tiles
 
         num_fgnd_tiles = np.count_nonzero(
             tile_fgnd_frac_list >= args.min_fgnd_frac)
@@ -181,7 +179,7 @@ def main(args):
         print('Number of foreground tiles = {0:d} ({1:2f}%%)'.format(
             num_fgnd_tiles, percent_fgnd_tiles))
 
-        print('Time taken = {}'.format(
+        print('Tile foreground fraction computation time = {}'.format(
             cli_utils.disp_time_hms(fgnd_frac_comp_time)))
 
     #
@@ -217,7 +215,7 @@ def main(args):
 
     print('Number of nuclei = {}'.format(len(nuclei_list)))
 
-    print('Time taken = {}'.format(
+    print('Nuclei detection time = {}'.format(
         cli_utils.disp_time_hms(nuclei_detection_time)))
 
     #
@@ -238,8 +236,8 @@ def main(args):
 
     total_time_taken = time.time() - total_start_time
 
-    print('Total analysis time = {}'.format(cli_utils.disp_time_hms(
-        total_time_taken)))
+    print('Total analysis time = {}'.format(
+        cli_utils.disp_time_hms(total_time_taken)))
 
 
 if __name__ == "__main__":