--- a/server/NucleiDetection/NucleiDetection.py
+++ b/server/NucleiDetection/NucleiDetection.py
@@ -167,7 +167,7 @@ def main(args):
 
         else:
 
-            tile_fgnd_frac_list = [1.0] * num_tiles
+            tile_fgnd_frac_list = np.full(num_tiles, 1.0)
 
         num_fgnd_tiles = np.count_nonzero(
             tile_fgnd_frac_list >= args.min_fgnd_frac)