--- a/server/cli_common/utils.py
+++ b/server/cli_common/utils.py
@@ -1,4 +1,3 @@
-import time
 from datetime import timedelta
 
 import numpy as np