--- a/server/cli_common/utils.py
+++ b/server/cli_common/utils.py
@@ -50,8 +50,9 @@ def segment_wsi_foreground_at_low_res(ts, lres_size=2048):
 
     # get image at low-res
     maxSize = max(ts_metadata['sizeX'], ts_metadata['sizeY'])
+    maxSize = float(min(maxSize, lres_size))
 
-    downsample_factor = 2 ** np.floor(np.log2(maxSize / lres_size))
+    downsample_factor = 2.0 ** np.floor(np.log2(maxSize / lres_size))
 
     fgnd_seg_mag = ts_metadata['magnification'] / downsample_factor